Rename: 폴더명 monitoring -> fc003 변경
