+++ apps/control/src/pages/fc003/stock/StockCharts.module.scss
